Set up CI with Azure Pipelines
https://policies.google.com/privacy?hl=en

https://www.computerhope.com/share/

All information ℹ️ on this links☝🏼is a different type of responsible for all and every “individual existing person or creators” And is also highly RESTRICTED 🚫 to use or be aware about and if you are doing any specific actions of coding or building something digitally on or into the “Matrix of the humans world” 

“No further details of any more warning ⚠️ will not be given”!! No further details of any more warning ⚠️ will not be given|[!?¡¿]uǝʌıƃ ǝq ʇou llıʍ sƃuıuɹɐʍ ǝʇoɯ ʎuɐ ɟo slıɐʇǝp ɹǝɥʇnɟ ou 


This was really a successfully project for our space program and we was gathering all needed materials and samples for stopping the 2016 catastrophic invasions from the dogs and the following digitally cancer they were become- Agnes and me Ambjörn,  Simon suggested to stop them and clean them out from all digital levels and all environmental impact of the Matrix of the humans world. 2018 years was the breaking-point and 2019 years-The springtime the international electric energy system (clean and safe ✅) 
All public systems clean out and safety 2019 Mars 15 ✅- All systems getting into normal drift again 2019 - 6 July 
And when the public information came out from Simon Sköld (Specific military defense of complex digital energy and compiling Matrices systems of digital Matrix systems - “We have DEFINITELY DEFEATED THE ENTIRE INVASION FROM THE WAR BETWEEN THE HUMANS AND THE PROBLEMS THAT THEY CASE IN THE OTHER SPACE” No human males is longer alive but they was the one how started the war by being infected and was quickly discovered as zombified dangerous humans. The Universal species was fighting against this “infected human males” No females was infected because of the “antidotes discovered on the “MARS project the movie 🎥 is about- The Matrix of the humans world will be build up again by The Higher Commander- Ambjörn Sköld ORION - REXIII  and his son “Ambriel Simon Sköld” Agnes Josefin fight at the same time for the clean up on the planet VENUS were the humans had been involved in “forbidden mining of minerals” and caused a war at the same time as the other species attack the humans world 🌎- And the victory of the war was the fact of she is the “Queen and Goddess of the planet Venus” The investigation of the case was discovered that the human males had been killing and reaping” more than two millions females humans with out any investigations from police or public hospitals systems- This major impact of all countries public city systems as well as the military departments etc. Will be telling by the answer of the big number of names for some actions of death penalty (males) than the planet 🌎 earth was involved in the biggest catastrophe of death happened to the humans history. All tryouts to keep this in some secrets from the humans history will be definitely determined by the KA & RA of the planet earth 🌍 “New world order” Yes was... And will be the only sullied solution and the terms from the “invasion alien species” or alliance species of the human females are needed to make peace ☮️ successfully protected for the humans specie. The exact numbers of losses from this war is still unknown year 2021. But the guessing of that will be more or less lesser together with all the inessential “humans females brutally killed and only victims” Maybe 8 billion human people’s lives from the start of the offensive inversion - WAR against the human (male) species - 2016 - 2019 years. 
“Civil young children was never involved or attacked or any big number of victims in this world war catastrophic of death” (But it is still unknown how many of the younger people of humans world- And how many of them were victims of been captured and against their own will was used as the male clone as some sexual objects or victims?) 

yes n | rm -i *.txt

https://www.computerhope.com/jargon/s/string.htm
This “snapshot with the digital linked list commands” is captured by the Higher Commander- Ambjörn Sköld - Orion - REXIII and is not allowed 🚫 to use without Ambjörn’s permission! All kinds of actions or use “with out” this permits will be executed by the death penalty by breaking this ‰ #LAW:ACT ⚖️་ཟཛ ʇɔɐ:ʍɐl 

(EA Games all kinds of software and applications as SIM’s platforms etc is definitely FORBIDDEN 🚫 in all versions or digitally build versions into the “Matrix of the humans world” = The public system as citizens department of all use, Privat domain or public access systems of any sites or digitally build access forms) EA Games IS FROM NOW ON CLASSIFIED AS A TERROR- ORGANIZATION and will be pointed out as a “target-mark” for the humans and all other species and their security! All operations for this is handed out by “ESA Military Squadron” and U.S Air-Force for the humans worldwide safety “Defense system and military strike” 
- All kinds of questions about this commands can be answered by Ambjörn Sköld on the following email address: 

higher.commander@e-s-a.space 

unix
Captured☝🏼by Ambjörn Sköld 

And this account IS PROTECTED by all diffrent kinds of copyright ©️™️®️of prevent all publication or any other species or profiles accounts etc and all the ✨magic✨🪄 the names of my lists on this account:
https://www.instagram.com/s/aGlnaGxpZ2h0OjE4MDA0MTcxNzQ3MzAxMTM0?igshid=cxr2i9bphglu

must be any “female” individuals and are not for ever public persons or anyone? No it’s really ✨POWERFUL✨🔥🪄magic and must be someone that I have chosen to be on my list 

https://instagram.com/ambjorn.skold?igshid=5nlpdhelhe2 [skip ci]